requirements.txt
absl-py==0.8.1
astor==0.8.0
cycler==0.10.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.3
h5py==2.10.0
joblib==0.14.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
mock==3.0.5
numpy==1.17.3
opencv-python==4.1.1.26
opt-einsum==3.1.0
pandas==0.25.2
Pillow==6.2.1
protobuf==3.10.0
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.3
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
tf-explain==0.1.0
tqdm==4.36.1
Werkzeug==0.16.0
wrapt==1.11.2
